Added comments to indicate csPoly2D::In() and csPoly3D::In() only works on convex polygons.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@33652 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/threading/thread.h
@@ -75,7 +75,7 @@ namespace Threading
   };
 
   /// OS specific thread identifier
-  typedef unsigned int ThreadID;
+  typedef unsigned long ThreadID;
 
 }
 }
